layout progress (whats next, read: tri-266)
"The layout works with 1 level of subflows, but after that it kinda breaks down. The issue is that you need the node sizes when you build the elk tree and do the layout, but we don't get access to those until we've laid everything out.

I think we'll have to do some sort of bottom-up/postorder stuff where we first lay out all the children, then get their calculated sizes and use that with laying out the parent"